test: adjust output of attribute tests
I classified the cases before adding the expression tests.
After checking the classified cases, I decided to write the output of the new expression tests in a future-oriented way, regardless of the output of the current version.
However, since the output of the attribute tests was written in a format that follows the output of the current version, these also needed to be adjusted in a future-oriented way for consistency reasons.